Change panic & error handling
Change the calls to `panic` to name the field that is expected. When `(*json.Decoder).Decode` returns `io.EOF`, return a publicly visible variable. That way the value can be checked for with `err == ErrRequestBodyIncomplete`. Also note that `io.EOF` doesn't mean that the request body from the client is empty, just that the request body did not contain a complete JSON object.